+++ src/presenter/routes/auth/auth_objects/login_request.rs
@@ -1,0 +1,9 @@
+use serde::Deserialize;
+
+#[derive(Deserialize, Debug, Clone)]
+pub struct LoginRequest<'a> {
+    #[serde(rename = "login")]
+    pub login: &'a str,
+    #[serde(rename = "password")]
+    pub password: &'a str,
+}
