+++ src/presenter/routes/auth/validators/jwt_decoder/decoder.rs
@@ -8,7 +8,10 @@ use hmac::{
 use jwt::{Header, Token, VerifyWithKey};
 use sha2::Sha256;
 
-use super::JwtDecoder;
+use super::{
+    objects::{JwtDecoderError, JwtDecoderResult},
+    JwtDecoder,
+};
 
 impl JwtDecoder for &str {
     fn decode_access(&self) -> Result<JwtDecoderResult, JwtDecoderError> {
@@ -83,43 +86,5 @@ impl JwtDecoder for &str {
             uuid: claims.get("uuid").unwrap().to_string(),
             username: claims.get("username").unwrap().to_string(),
         })
-    }
-}
-
-pub struct JwtDecoderResult {
-    pub uuid: String,
-    pub username: String,
-}
-
-#[derive(Debug)]
-pub enum JwtDecoderError {
-    InvalidEnvSecret,
-    InvalidSignature,
-    ExpiredSignature,
-    ParceError(String),
-}
-
-impl std::fmt::Display for JwtDecoderError {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        match self {
-            JwtDecoderError::InvalidEnvSecret => write!(f, "Invalid env secret"),
-            JwtDecoderError::InvalidSignature => write!(f, "Invalid signature"),
-            JwtDecoderError::ExpiredSignature => write!(f, "Expired signature"),
-            JwtDecoderError::ParceError(message) => write!(f, "Parce error: {}", message),
-        }
-    }
-}
-
-impl PartialEq for JwtDecoderError {
-    fn eq(&self, other: &Self) -> bool {
-        match (self, other) {
-            (JwtDecoderError::InvalidEnvSecret, JwtDecoderError::InvalidEnvSecret) => true,
-            (JwtDecoderError::InvalidSignature, JwtDecoderError::InvalidSignature) => true,
-            (JwtDecoderError::ExpiredSignature, JwtDecoderError::ExpiredSignature) => true,
-            (JwtDecoderError::ParceError(message1), JwtDecoderError::ParceError(message2)) => {
-                message1 == message2
-            }
-            _ => false,
-        }
     }
 }
