--- src/utils/jwt_util/tests/mod.rs
@@ -1,2 +1,0 @@
-mod mapper;
-mod test_decoder;
